Fix URL generation for RSS feed
RSS feed generation was using `base-directory' to determine the place
where the blog was going to be published, this would turn into
something like `~/org-blog/blog-2012-05-06-1041.org', which isn't
where other people can browse to it.

Now a `:blog-url' option must be added to the
`org-publish-project-alist' project for the blog which points to the
base URL for the blog.